fix: fix generation of EF tar file
