+++ packages/edge-bundler/node/formats/tarball.ts
@@ -9,8 +9,9 @@ import { DenoBridge } from '../bridge.js'
 import { Bundle, BundleFormat } from '../bundle.js'
 import { EdgeFunction } from '../edge_function.js'
 import { FeatureFlags } from '../feature_flags.js'
+import { listRecursively } from '../utils/fs.js'
 import { ImportMap } from '../import_map.js'
-import { getDirectoryHash, getStringHash } from '../utils/sha256.js'
+import { getFileHash } from '../utils/sha256.js'
 
 const TARBALL_EXTENSION = '.tar'
 
@@ -42,8 +43,8 @@ export const bundle = async ({
   importMap,
   vendorDirectory,
 }: BundleTarballOptions): Promise<Bundle> => {
-  const sideFilesDir = await tmp.dir({ unsafeCleanup: true })
-  const cleanup = [sideFilesDir.cleanup]
+  const bundleDir = await tmp.dir({ unsafeCleanup: true })
+  const cleanup = [bundleDir.cleanup]
 
   let denoDir = vendorDirectory ? path.join(vendorDirectory, 'deno_dir') : undefined
 
@@ -87,7 +88,7 @@ export const bundle = async ({
       '--quiet',
       '--code-splitting',
       '--outdir',
-      distDirectory,
+      bundleDir.path,
       ...functions.map((func) => func.path),
     ],
     {
@@ -95,63 +96,39 @@ export const bundle = async ({
     },
   )
 
-  const manifestPath = path.join(sideFilesDir.path, 'manifest.json')
+  const manifestPath = path.join(bundleDir.path, '___netlify-edge-functions.json')
   const manifestContents = JSON.stringify(manifest)
   await fs.writeFile(manifestPath, manifestContents)
 
-  const denoConfigPath = path.join(sideFilesDir.path, 'deno.json')
+  const denoConfigPath = path.join(bundleDir.path, 'deno.json')
   const denoConfigContents = JSON.stringify(importMap.getContentsWithRelativePaths())
   await fs.writeFile(denoConfigPath, denoConfigContents)
 
-  const rootLevel = await fs.readdir(distDirectory)
-  const hash = await getDirectoryHash(distDirectory)
   const tarballPath = path.join(distDirectory, buildID + TARBALL_EXTENSION)
-
   await fs.mkdir(path.dirname(tarballPath), { recursive: true })
 
-  // Adding all the bundled files.
   await tar.create(
     {
-      cwd: distDirectory,
+      cwd: bundleDir.path,
       file: tarballPath,
+      noDirRecurse: true,
       onWriteEntry(entry) {
-        entry.path = getUnixPath(`./${entry.path}`)
+        const relativePath = path.relative(bundleDir.path, path.join('/', entry.path))
+        const normalizedPath = getUnixPath(relativePath)
+
+        entry.path = normalizedPath
       },
     },
-    rootLevel,
+    await listRecursively(bundleDir.path),
   )
 
-  // Adding `deno.json`.
-  await tar.update(
-    {
-      cwd: distDirectory,
-      file: tarballPath,
-      onWriteEntry(entry) {
-        entry.path = './deno.json'
-      },
-    },
-    [denoConfigPath],
-  )
+  const hash = await getFileHash(tarballPath)
 
-  // Adding the manifest file.
-  await tar.update(
-    {
-      cwd: distDirectory,
-      file: tarballPath,
-      onWriteEntry(entry) {
-        entry.path = './___netlify-edge-functions.json'
-      },
-    },
-    [manifestPath],
-  )
-
-  await Promise.all(cleanup)
-
-  const finalHash = [hash, getStringHash(manifestContents), getStringHash(denoConfigContents)].join('')
+  await Promise.allSettled(cleanup)
 
   return {
     extension: TARBALL_EXTENSION,
     format: BundleFormat.TARBALL,
-    hash: finalHash,
+    hash,
   }
 }
